Split out separate sockets for server to respond to discovery requests on - now that sendingSockets are in multicast group, it no longer works to use these to send to arbitrary addresses.
Move discovery on client to separate one-shot process; since we only use the sendingSocket in client for discovery, there's no need to explicitly run it
before starting to listen on other sockets - the data received on each will never be mixed.

Store in instvars data needed for startUp of clients - since we pass in socket to process creation, we need to store the other parameters as well so we
can recreate on valid sockets after image startup.

Fix byebyeResponse to take multicastinfo parameter as well.