Only store interfaces we could setup sockets on in runningOn.
Change the address lookup once more; always consult local domain.
Experienced the case that when getHostName contains a domain assigned by primary network, and there are connections to other networks with a different domain,  the other networks would not be returned (on OSX). Cross fingers and hope it works a bit better on other os's as well.