If an image is running on a machine that is waking up from sleep mode, udp sockets may be invalid (at least on OSX), and fail to send/receive traffic. The cache expulsion process will then immediately fail it's timeout check, and try to rediscover the service, at which point the socket will raise a primitiveFailure trying to send  data; we can then handle this by restarting the client, creating new sockets and resetting the cache.
If waking a computer with an image where the client had yet to discover an available service, we're currently SOL though; initial discovery will already have run, and the client be stuck listening to change notifications on sockets that will never receive traffic.

Also remember to run unavailability handler on currently known services when client is shut down.